feat(dialog): track controllers in service
Dialog controllers were not tracked or destroyed by the service. Thus, there was no way to determine whether a dialog was active. We add a controllers array to the service to track all active dialog controllers, adding them upon dialog initialization and removing their reference when the `open` promise is resolved. This does not remove the collection from the renderer - it merely separates the concerns.

Fixes #121.