fix(dialog-controller): allow default DI injection
The DialogController constructor assumed a DialogRenderer instance to assign settings. This adds a check to prevent an error if the renderer does not have that property or if the renderer is undefined. This should allow standard DI injection without having to declare a renderer in the container, so that the dialogController can possibly be used for other purposes.

Fixes #157.